11/10 2nd upload
I still don't fully understand how to declutter my homepage. I put in some styles like fonts, box sizes, and all that. I can sort of get the fonts to work, the box sizes are not working thought. I also tried changing the submit box to red but the color was only applied to the text.