Not really many changes from the one earlier
I just resectioned out the home. I added a bunch to the CSS. That is my main concern with the glitchiness of this. I'm trying to get everything in place on the home. I feel like it may be my width and height stuff.  I also tried to style the bioimage and couldn't get any reaction.